local changes

+++ src/app/webapp-common/projects/containers/projects-page/common-projects-page.component.ts
@@ -1,4 +1,4 @@
-import {Component, inject, OnDestroy, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {Store} from '@ngrx/store';
 import {
   CommonProjectReadyForDeletion,
@@ -13,18 +13,19 @@ import {
   getAllProjectsPageProjects,
   resetProjects,
   resetProjectsSearchQuery,
-  resetReadyToDelete, setCurrentScrollId,
+  resetReadyToDelete,
   setProjectsOrderBy,
   setProjectsSearchQuery,
   updateProject
 } from '../../common-projects.actions';
-import {ActivatedRoute, Params, Router} from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 import {MatDialog, MatDialogRef} from '@angular/material/dialog';
 import {ProjectsGetAllResponseSingle} from '~/business-logic/model/projects/projectsGetAllResponseSingle';
 import {ProjectDialogComponent} from '@common/shared/project-dialog/project-dialog.component';
 import {combineLatest, Observable, Subscription} from 'rxjs';
 import {
   debounceTime,
+  distinctUntilChanged,
   distinctUntilKeyChanged,
   filter,
   map,
@@ -34,7 +35,7 @@ import {
   withLatestFrom
 } from 'rxjs/operators';
 import {ConfirmDialogComponent} from '@common/shared/ui-components/overlay/confirm-dialog/confirm-dialog.component';
-import * as coreProjectsActions from '@common/core/actions/projects.actions';
+import * as coreProjectsActions from '../../../core/actions/projects.actions';
 import {setDeep, setSelectedProjectId} from '@common/core/actions/projects.actions';
 import {initSearch, resetSearch} from '@common/common-search/common-search.actions';
 import {SearchState, selectSearchQuery} from '@common/common-search/common-search.reducer';
@@ -44,7 +45,7 @@ import {
   popupEntitiesListConst,
   readyForDeletionFilter
 } from '~/features/projects/projects-page.utils';
-import {selectRouterConfig, selectRouterParams} from '@common/core/reducers/router-reducer';
+import {selectRouterParams} from '@common/core/reducers/router-reducer';
 import {selectShowOnlyUserWork} from '@common/core/reducers/users-reducer';
 import {Project} from '~/business-logic/model/projects/project';
 import {CommonDeleteDialogComponent} from '@common/shared/entity-page/entity-delete/common-delete-dialog.component';
@@ -53,6 +54,7 @@ import {isExample} from '@common/shared/utils/shared-utils';
 import {selectSelectedProject} from '@common/core/reducers/projects.reducer';
 import {selectActiveWorkspaceReady} from '~/core/reducers/view.reducer';
 import {EntityTypeEnum} from '~/shared/constants/non-common-consts';
+import {selectIsDatasets} from '@common/experiments-compare/reducers';
 
 @Component({
   selector: 'sm-common-projects-page',
@@ -90,32 +92,37 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
   public selectedProjectId$: Observable<string>;
   public allExamples: boolean;
   /* eslint-enable @typescript-eslint/naming-convention */
-  protected readonly searchQuery$: Observable<SearchState['searchQuery']>;
+  private projectReadyForDeletion$: Observable<CommonProjectReadyForDeletion>;
+  private readonly searchQuery$: Observable<SearchState['searchQuery']>;
   private projectDialog: MatDialogRef<ProjectDialogComponent, any>;
-  public selectedProject$: Observable<Project>;
+  private selectedProject$: Observable<Project>;
   public projectId: string;
-  public subs = new Subscription();
+  private subs = new Subscription();
   private selectedProject: Project;
-  protected store = inject(Store);
-  protected router = inject(Router);
-  protected route = inject(ActivatedRoute);
-  protected dialog = inject(MatDialog);
 
-  constructor() {
+  constructor(
+    protected store: Store<any>,
+    protected router: Router,
+    protected route: ActivatedRoute,
+    protected dialog: MatDialog
+  ) {
     this.searchQuery$ = this.store.select(selectSearchQuery);
     this.projectsOrderBy$ = this.store.select(selectProjectsOrderBy);
     this.projectsSortOrder$ = this.store.select(selectProjectsSortOrder);
     this.noMoreProjects$ = this.store.select(selectNoMoreProjects);
-    this.selectedProjectId$ = this.store.select(selectRouterParams).pipe(map((params: Params) => params?.projectId));
+    this.selectedProjectId$ = this.store.select(selectRouterParams).pipe(map(params => params?.projectId));
     this.selectedProject$ = this.store.select(selectSelectedProject).pipe(tap(selectedProject => this.selectedProject = selectedProject));
+    this.projectReadyForDeletion$ = this.store.select(selectProjectReadyForDeletion).pipe(
+      distinctUntilChanged(),
+      filter(readyForDeletion => readyForDeletionFilter(readyForDeletion)));
 
     this.projectsList$ = combineLatest([
       this.store.select(selectProjects),
       this.store.select(selectSelectedProject)
     ]).pipe(
-      debounceTime(50),
-      withLatestFrom(this.selectedProjectId$, this.searchQuery$, this.store.select(selectRouterConfig)),
-      map(([[projectsList, selectedProject = {} as Project], selectedProjectId, searchQuery, config]) => {
+      debounceTime(0),
+      withLatestFrom(this.selectedProjectId$, this.searchQuery$, this.store.select(selectIsDatasets)),
+      map(([[projectsList, selectedProject], selectedProjectId, searchQuery, isDatasets]) => {
         this.searching = searchQuery?.query.length > 0;
         this.allExamples = projectsList?.length > 0 && projectsList?.every(project => isExample(project));
         if (projectsList === null) {
@@ -124,8 +131,8 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
         if ((searchQuery?.query || searchQuery?.regExp)) {
           return projectsList;
         } else {
-          if ((selectedProject?.sub_projects?.length === 0 || this.shouldReRoute(selectedProject, config)) && selectedProjectId === selectedProject?.id) {
-            this.noProjectsReRoute();
+          if (selectedProject?.sub_projects?.length === 0 && selectedProjectId === selectedProject?.id) {
+            this.router.navigate(['..', 'experiments'], {relativeTo: isDatasets ? this.route : this.route.parent});
             return [];
           }
           const pageProjectsList = this.getExtraProjects(selectedProjectId, selectedProject);
@@ -133,41 +140,34 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
         }
       }),
     );
+
     this.syncAppSearch();
   }
 
-  noProjectsReRoute() {
-    return this.router.navigate(['..', 'experiments'], {relativeTo: this.route.parent});
-  }
-
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  shouldReRoute(selectedProject, config) {
-    return false;
-  };
-
   protected getExtraProjects(selectedProjectId, selectedProject) {
     return [{
-      ...((selectedProjectId && selectedProject?.id) ? selectedProject : this.ALL_EXPERIMENTS_CARD),
+      ...((selectedProjectId && selectedProject) ? selectedProject : this.ALL_EXPERIMENTS_CARD),
       // eslint-disable-next-line @typescript-eslint/naming-convention
-      id: selectedProjectId ? selectedProjectId : '*',
-      name: 'All Experiments',
-      basename: 'All Experiments',
-      // eslint-disable-next-line @typescript-eslint/naming-convention
-      sub_projects: null,
+      id: selectedProjectId ? selectedProjectId : '*', name: 'All Experiments', sub_projects: null,
     } as ProjectsGetAllResponseSingle];
   }
 
   ngOnInit() {
-    this.store.dispatch(resetProjects());
+    // this.store.dispatch(new ResetSelectedProject());
     this.store.dispatch(setDeep({deep: false}));
-    // Todo: join the 2 subscriptions to 1 selector.
     this.subs.add(this.store.select(selectActiveWorkspaceReady)
-      .pipe(filter(ready => !!ready), take(1))
-      .subscribe(() => this.store.dispatch(getAllProjectsPageProjects())));
+      .pipe(filter(ready => ready), take(1))
+      .subscribe(() => {
+        this.store.dispatch(getAllProjectsPageProjects());
+      }));
 
     this.subs.add(this.selectedProjectId$.pipe(
-    ).subscribe((projectId) => {
-      this.projectId = projectId;
+      tap(projectId => this.projectId = projectId),
+      filter(projectId => !projectId),
+      distinctUntilChanged(),
+    ).subscribe(() => {
+        this.store.dispatch(resetProjectsSearchQuery());
+        this.store.dispatch(getAllProjectsPageProjects());
       }
     ));
 
@@ -176,54 +176,27 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
       distinctUntilKeyChanged('id'),
       skip(1),
     ).subscribe(() => {
-      this.store.dispatch(setCurrentScrollId({scrollId: null}));
+      this.store.dispatch(resetProjectsSearchQuery());
       this.store.dispatch(getAllProjectsPageProjects());
     }));
     this.subs.add(this.store.select(selectShowOnlyUserWork)
       .pipe(skip(1))
       .subscribe(() => {
-        this.store.dispatch(resetProjects());
+        this.store.dispatch(resetProjectsSearchQuery());
         this.store.dispatch(getAllProjectsPageProjects());
       }));
 
-    this.subs.add(this.store.select(selectProjectReadyForDeletion)
-      .pipe(
-        filter(readyForDeletion => readyForDeletion !== null && readyForDeletionFilter(readyForDeletion))
-      )
-      .subscribe(readyForDeletion => {
-        if (isDeletableProject(readyForDeletion)) {
-          this.showDeleteDialog(readyForDeletion);
-        } else {
-          this.showConfirmDialog(readyForDeletion);
-        }
-      })
-    );
-    this.setupBreadcrumbsOptions();
+    this.subs.add(this.projectReadyForDeletion$.subscribe(readyForDeletion => {
+      if (isDeletableProject(readyForDeletion)) {
+        this.showDeleteDialog(readyForDeletion);
+      } else {
+        this.showConfirmDialog(readyForDeletion);
+      }
+    }));
   }
 
   protected getDeletePopupEntitiesList() {
     return 'experiment';
-  }
-
-  setupBreadcrumbsOptions(){
-    this.subs.add(this.selectedProject$.pipe().subscribe((selectedProject) => {
-      this.store.dispatch(coreProjectsActions.setBreadcrumbsOptions({
-        breadcrumbOptions: {
-          showProjects: !!selectedProject,
-          featureBreadcrumb: {
-            name: 'PROJECTS',
-            url: 'projects'
-          },
-          projectsOptions: {
-            basePath: 'projects',
-            filterBaseNameWith: null,
-            compareModule: null,
-            showSelectedProject: true,
-            ...(selectedProject && {selectedProjectBreadcrumb: {name: selectedProject?.basename}})
-          }
-        }
-      }));
-    }));
   }
 
   private showConfirmDialog(readyForDeletion: CommonProjectReadyForDeletion) {
@@ -247,8 +220,8 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
     });
   }
 
-  private showDeleteDialog(readyForDeletion) {
-    this.dialog.open(CommonDeleteDialogComponent, {
+  private showDeleteDialog(readyForDeletion: CommonProjectReadyForDeletion) {
+    const confirmDialogRef = this.dialog.open(CommonDeleteDialogComponent, {
       data: {
         entity: readyForDeletion.project,
         numSelected: 1,
@@ -257,13 +230,13 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
       },
       width: '600px',
       disableClose: true
-    }).afterClosed()
-      .subscribe((confirmed) => {
+    });
+    confirmDialogRef.afterClosed().subscribe((confirmed) => {
       if (confirmed) {
+        this.store.dispatch(resetDeleteState());
         this.store.dispatch(resetProjects());
         this.store.dispatch(getAllProjectsPageProjects());
       }
-      this.store.dispatch(resetDeleteState());
     });
   }
 
@@ -272,7 +245,6 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
     this.subs.unsubscribe();
     this.store.dispatch(resetReadyToDelete());
     this.store.dispatch(resetProjectsSearchQuery());
-    this.store.dispatch(resetProjects());
   }
 
   stopSyncSearch() {
@@ -282,8 +254,6 @@ export class CommonProjectsPageComponent implements OnInit, OnDestroy {
 
   syncAppSearch() {
     this.store.dispatch(initSearch({payload: `Search for ${this.getName()}s`}));
-    this.subs.add(this.searchQuery$.pipe(take(1)).subscribe(query =>
-      this.store.dispatch(setProjectsSearchQuery({...query, skipGetAll: true}))));
     this.subs.add(this.searchQuery$.pipe(skip(1)).subscribe(query => this.search(query)));
   }
